Run cppcheck and fix errors
