+++ sensing/preprocessor/pointcloud/pointcloud_preprocessor/include/pointcloud_preprocessor/passthrough_filter/passthrough_filter_uint16_nodelet.hpp
@@ -2,7 +2,7 @@
 
 #include "pcl/search/pcl_search.h"
 #include "pointcloud_preprocessor/filter.hpp"
-#include "pointcloud_preprocessor/passthrough_filter/passthrough_uint16.h"
+#include "pointcloud_preprocessor/passthrough_filter/passthrough_uint16.hpp"
 
 namespace pointcloud_preprocessor
 {
