+++ sensing/preprocessor/pointcloud/pointcloud_preprocessor/src/passthrough_filter/passthrough_uint16.cpp
@@ -38,7 +38,7 @@
  *
  */
 
-#include "pointcloud_preprocessor/passthrough_filter/passthrough_uint16.hpp"
+#include "pointcloud_preprocessor/passthrough_filter/passthrough_uint16_impl.hpp"
 
 //////////////////////////////////////////////////////////////////////////
 void pcl::PassThroughUInt16<pcl::PCLPointCloud2>::applyFilter(PCLPointCloud2 & output)
